% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version.R
\name{gdalbin}
\alias{gdalbin}
\alias{vapour_version}
\alias{rgdal_version}
\alias{sf_version}
\title{Get version}
\usage{
vapour_version()

rgdal_version()

sf_version()
}
\value{
character string
}
\description{
Get version of GDAL
}
\details{
Prints version
}
\examples{
vapour_version()
rgdal_version()
}
